Moved archive events page to app directory
