+++ app/utils/countryFlags.ts
@@ -1,0 +1,12 @@
+import getUnicodeFlagIcon from "country-flag-icons/unicode";
+
+export const getCountryFlag = (nationality: string) => {
+  // i18n-nationality has Monaco as "Monacan" instead of "Monegasque"
+  if (nationality === "Monegasque") nationality = "Monacan";
+
+  const nationalities = require("i18n-nationality");
+  nationalities.registerLocale(require("i18n-nationality/langs/en.json"));
+  const countryCode = nationalities.getAlpha2Code(nationality, "en");
+  if (!countryCode) return "N/A"; // Not recognised by i18n-nationality library
+  return getUnicodeFlagIcon(countryCode);
+};
